// src/app/interfaces/antecedentes.ts
export interface Antecedentes {
  idantecedente?: number;
  idpaciente?: number;
  parabeletza?: number;
  a_diabetes?: number;
  a_gastritis?: number;
  a_hipertension?: number;
  a_ansiedad?: number;
  a_operaciones?: number;
  a_hormonales?: number;
  a_cesarea?: number;
  a_anticonceptivos?: number;
  a_cigarro_alcohol?: number;
  a_otros?: number;
  creacion?: string;
  modificacion?: string;
}